requirements.txt
streamlit==1.27.2
google-generativeai==0.3.1
PyPDF2==3.0.1
python-docx==0.8.11
pandas==2.1.1
openpyxl==3.1.2
langchain==0.0.352
faiss-cpu==1.7.4
sentence-transformers==2.2.2
google-api-python-client==2.108.0
protobuf==4.25.1